A EngineGraphQLGitHub class

A class for methods wrapping GitHub's GraphQL API responses.

Super class

GitStats::EngineGraphQL -> EngineGraphQLGitHub

Methods

Public methods

EngineGraphQLGitHub$new()

EngineGraphQLGitHub$pull_orgs()

EngineGraphQLGitHub$pull_repos()

EngineGraphQLGitHub$pull_repos_supportive()

EngineGraphQLGitHub$pull_commits()

EngineGraphQLGitHub$pull_commits_supportive()

EngineGraphQLGitHub$clone()

Inherited methods

GitStats::EngineGraphQL$gql_response()

GitStats::EngineGraphQL$pull_users()

Method new()

Create EngineGraphQLGitHub object.

Usage

EngineGraphQLGitHub$new(gql_api_url, token, scan_all = FALSE)

Arguments

gql_api_url

GraphQL API url.

token

A token.

scan_all

A boolean.

Method pull_orgs()

Get all orgs from GitHub.

Usage

EngineGraphQLGitHub$pull_orgs()

Method pull_repos()

A method to retrieve all repositories for an organization in a table format.

Usage

EngineGraphQLGitHub$pull_repos(org, settings)

Arguments

org

An organization.

settings

A list of GitStats settings.

Returns

A table.

Method pull_repos_supportive()

An empty method to satisfy engine iterator.

Usage

EngineGraphQLGitHub$pull_repos_supportive(org, settings)

Arguments

org

An organization.

settings

A list of GitStats settings.

Returns

Nothing.

Method pull_commits()

Method to pull all commits from organization, optionally filtered by team members.

Usage

EngineGraphQLGitHub$pull_commits(org, date_from, date_until, settings)

Arguments

org

An organization.

date_from

A starting date to look commits for.

date_until

An end date to look commits for.

settings

A list of GitStats settings.

Returns

A table of commits.

Method pull_commits_supportive()

Method to get commits.

Usage

EngineGraphQLGitHub$pull_commits_supportive(
  org,
  date_from,
  date_until = Sys.date(),
  settings
)

Arguments

org

An organization.

date_from

A starting date to look commits for.

date_until

An end date to look commits for.

settings

A list of GitStats settings.

Details

This method must exist as it is called from the GitHost wrapper above.

Returns

A table of commits.

Method clone()

The objects of this class are cloneable with this method.

Usage

EngineGraphQLGitHub$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.