A EngineRestGitLab class

A class for methods wrapping GitLab's REST API responses.

Super class

GitStats::EngineRest -> EngineRestGitLab

Methods

Public methods

EngineRestGitLab$pull_repos()

EngineRestGitLab$pull_repos_supportive()

EngineRestGitLab$pull_repos_contributors()

EngineRestGitLab$pull_commits()

EngineRestGitLab$clone()

Inherited methods

GitStats::EngineRest$check_organizations()

GitStats::EngineRest$initialize()

GitStats::EngineRest$response()

Method pull_repos()

A method to retrieve all repositories for an organization in a table format.

Usage

EngineRestGitLab$pull_repos(org, settings)

Arguments

org

A character, a group of projects.

settings

A list of GitStats settings.

Returns

A table.

Method pull_repos_supportive()

An empty method to satisfy engine iterator.

Usage

EngineRestGitLab$pull_repos_supportive(org, settings)

Arguments

org

An organization.

settings

A list of GitStats settings.

Returns

Nothing.

Method pull_repos_contributors()

A method to add information on repository contributors.

Usage

EngineRestGitLab$pull_repos_contributors(repos_table)

Arguments

repos_table

A table of repositories.

Returns

A table of repositories with added information on contributors.

Method pull_commits()

GitLab private method to derive commits from repo with REST API.

Usage

EngineRestGitLab$pull_commits(
  org,
  date_from,
  date_until = Sys.date(),
  settings
)

Arguments

org

A character, a group of projects.

date_from

A starting date to look commits for.

date_until

An end date to look commits for.

settings

A list of GitStats settings.

Returns

A table of commits.

Method clone()

The objects of this class are cloneable with this method.

Usage

EngineRestGitLab$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.