A GitHost superclass

A GitHost superclass

A GitHost superclass

Methods

Public methods

GitHost$new()

GitHost$pull_repos()

GitHost$pull_repos_contributors()

GitHost$pull_commits()

GitHost$pull_users()

GitHost$clone()

Method new()

Create a new GitHost object

Usage

GitHost$new(orgs = NA, token = NA, api_url = NA)

Arguments

orgs

A character vector of organisations (owners of repositories in case of GitHub and groups of projects in case of GitLab).

token

A token.

api_url

An API url.

Returns

A new GitHost object

Method pull_repos()

A method to list all repositories for an organization, a team or by a keyword.

Usage

GitHost$pull_repos(settings, add_contributors = FALSE)

Arguments

settings

A list of GitStats settings.

add_contributors

A boolean to decide whether to add contributors column to repositories table.

Returns

A data.frame of repositories.

Method pull_repos_contributors()

A method to add information on repository contributors.

Usage

GitHost$pull_repos_contributors(repos_table)

Arguments

repos_table

A table of repositories.

Returns

A table of repositories with added information on contributors.

Method pull_commits()

A method to get information on commits.

Usage

GitHost$pull_commits(date_from, date_until = Sys.Date(), settings)

Arguments

date_from

A starting date to look commits for.

date_until

An end date to look commits for.

settings

A list of GitStats settings.

Returns

A data.frame of commits

Method pull_users()

Get information about users

Usage

GitHost$pull_users(users)

Arguments

users

A character vector of users

Returns

Table of users

Method clone()

The objects of this class are cloneable with this method.

Usage

GitHost$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.